[refactor] Breaking out key codes
Trying to start breaking apart the large circular dependencies between
modules. Most of the dependencies are similar to this pattern where
there is one stand-alone member that is needed, so the entire module
gets brought along with.
