+++ utils/helpers/keyCodes.ts
@@ -1,0 +1,30 @@
+import { IMap } from "../../src/types";
+
+/**
+ * Keyboard key codes
+ */
+export const KEY: IMap<number> = {
+    A:              65,
+    B:              66,
+    C:              67,
+    CTRL:           17,
+    D:              68,
+    DOWNARROW:      40,
+    E:              69,
+    ENTER:          13,
+    F:              70,
+    H:              72,
+    J:              74,
+    K:              75,
+    L:              76,
+    M:              77,
+    N:              78,
+    O:              79,
+    P:              80,
+    R:              82,
+    S:              83,
+    TAB:            9,
+    U:              85,
+    UPARROW:        38,
+    W:              87,
+};
